+++ pkg/aws/v1alpha3/awscontrolplane/validate_awscontrolplane_test.go
@@ -2,9 +2,11 @@ package awscontrolplane
 
 import (
 	"context"
+	"fmt"
 	"strconv"
 	"testing"
 
+	"github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/giantswarm/micrologger/microloggertest"
 
 	unittest "github.com/giantswarm/aws-admission-controller/v3/pkg/unittest/v1alpha3"
@@ -274,7 +276,7 @@ func TestAZUnique(t *testing.T) {
 				logger:                 microloggertest.New(),
 			}
 
-			admissionRequest, err := unittest.CustomAdmissionRequestAWSControlPlane(tc.chosenAZs)
+			admissionRequest, err := unittest.CustomAdmissionRequestAWSControlPlane(nil, tc.chosenAZs)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -393,6 +395,81 @@ func TestInstanceTypeValid(t *testing.T) {
 			}
 
 			allowed, _ := validate.Validate(&admissionRequest)
+			if allowed != tc.allowed {
+				t.Fatalf("expected %v to not to differ from %v", allowed, tc.allowed)
+			}
+		})
+	}
+}
+
+func TestAnnotations(t *testing.T) {
+	testCases := []struct {
+		ctx  context.Context
+		name string
+
+		allowed     bool
+		annotations map[string]string
+	}{
+		{
+			ctx:  context.Background(),
+			name: "case 0",
+
+			allowed:     true,
+			annotations: map[string]string{annotation.AWSEBSVolumeIops: "16000", annotation.AWSEBSVolumeThroughput: "1000"},
+		},
+		{
+			ctx:  context.Background(),
+			name: "case 1",
+
+			allowed:     false,
+			annotations: map[string]string{annotation.AWSEBSVolumeIops: "160", annotation.AWSEBSVolumeThroughput: "1000"},
+		},
+		{
+			ctx:  context.Background(),
+			name: "case 2",
+
+			allowed:     false,
+			annotations: map[string]string{annotation.AWSEBSVolumeIops: "3000", annotation.AWSEBSVolumeThroughput: "2000"},
+		},
+		{
+			ctx:  context.Background(),
+			name: "case 3",
+
+			allowed:     false,
+			annotations: map[string]string{annotation.AWSEBSVolumeIops: "5000.40", annotation.AWSEBSVolumeThroughput: "1000"},
+		},
+		{
+			ctx:  context.Background(),
+			name: "case 4",
+
+			allowed: true,
+		},
+	}
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			var err error
+
+			fakeK8sClient := unittest.FakeK8sClient()
+			validate := &Validator{
+				validAvailabilityZones: unittest.DefaultAvailabilityZones(),
+				validInstanceTypes:     unittest.DefaultInstanceTypes(),
+				k8sClient:              fakeK8sClient,
+				logger:                 microloggertest.New(),
+			}
+
+			admissionRequest, err := unittest.CustomAdmissionRequestAWSControlPlane(tc.annotations, unittest.DefaultAvailabilityZones())
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			organization := unittest.DefaultOrganization()
+			err = fakeK8sClient.CtrlClient().Create(tc.ctx, organization)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			allowed, err := validate.Validate(&admissionRequest)
+			fmt.Print(err)
 			if allowed != tc.allowed {
 				t.Fatalf("expected %v to not to differ from %v", allowed, tc.allowed)
 			}