+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validate EBS volume iops and throughput settings.
+
 ## [3.6.3] - 2021-12-17
 
 ### Fixed